Moved code from FormMediaDownloader to ServerFormDownloaderUseCases
