+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/ServerFormDownloader.java
@@ -93,8 +93,7 @@ public class ServerFormDownloader implements FormDownloader {
 
             // download media files if there are any
             if (fd.getManifest() != null && !fd.getManifest().getMediaFiles().isEmpty()) {
-                FormMediaDownloader mediaDownloader = new FormMediaDownloader(formsRepository, formSource);
-                newAttachmentsDetected = mediaDownloader.download(fd, tempMediaPath, tempDir, stateListener);
+                newAttachmentsDetected = ServerFormDownloaderUseCases.download(formsRepository, formSource, fd, tempMediaPath, tempDir, stateListener);
             }
 
             ServerFormDownloaderUseCases.copySavedFileFromPreviousFormVersionIfExists(formsRepository, fd.getFormId(), tempMediaPath);
